refactor: ATLAS-NONE: Move metadata converters to more appropriate namespace.
